`processLinkEvents` in `DrivesVehicle` on `Driving` state when reached the end
