+++ src/main/scala/beam/agentsim/agents/PersonAgent.scala
@@ -203,43 +203,6 @@ class PersonAgent(
     else Right(_experiencedBeamPlan.activities(ind))
   }
 
-  def getAverageTravelTime(linkTravelTime: IndexedSeq[Int]):IndexedSeq[Int] = {
-    if (linkTravelTime.isEmpty) Vector()
-    else {
-      // FIXME There are the way to make it faster and even lazy
-      // We don't need the travel time for the last link, so we drop it (dropRight(1))
-      // We average travel time for the first link
-      linkTravelTime.zipWithIndex.map { case (travelTime, idx) =>
-        if (idx == 0) {
-          travelTime / 2
-        }
-        else {
-          travelTime
-        }
-      }
-    }
-  }
-
-  def processLinkEvents(itineraries: Seq[EmbodiedBeamTrip]): Unit = {
-    itineraries.foreach { itinerary =>
-     itinerary.legs.foreach { leg =>
-        val vehicleId = leg.beamVehicleId
-        val path = leg.beamLeg.travelPath
-        if (path.linkTravelTime.nonEmpty) {
-          // FIXME once done with debugging, make this code faster
-          val avgTravelTimeWithoutLast = getAverageTravelTime(path.linkTravelTime).dropRight(1)
-          val links = path.linkIds
-          val linksWithTime = links.sliding(2).zip(avgTravelTimeWithoutLast.iterator)
-          linksWithTime.foreach {
-            case (Seq(from, to), timeAtNode) =>
-              eventsManager.processEvent(new LinkLeaveEvent(timeAtNode, vehicleId, Id.createLinkId(from)))
-              eventsManager.processEvent(new LinkEnterEvent(timeAtNode, vehicleId, Id.createLinkId(to)))
-          }
-        }
-      }
-    }
-  }
-
   def currentActivity(data: BasePersonData): Activity =
     _experiencedBeamPlan.activities(data.currentActivityIndex)
